requirements.txt
PyQt5
requests
python-dotenv
pywifi
comtypes
